# Conflicts:
#	src/main/java/com/epam/cwlhub/exceptions/unchecked/UserException.java
#	src/main/java/com/epam/cwlhub/services/impl/UserServiceImpl.java
#	src/main/java/com/epam/cwlhub/servlets/LoginServlet.java

+++ src/main/java/com/epam/cwlhub/services/impl/GroupServiceImpl.java
@@ -3,11 +3,10 @@ package com.epam.cwlhub.services.impl;
 import com.epam.cwlhub.dao.GroupDao;
 import com.epam.cwlhub.dao.impl.GroupDaoImpl;
 import com.epam.cwlhub.entities.group.Group;
-import com.epam.cwlhub.entities.user.UserEntity;
+import com.epam.cwlhub.exceptions.unchecked.GroupException;
 import com.epam.cwlhub.services.GroupService;
 
 import java.util.List;
-import java.util.Optional;
 
 public class GroupServiceImpl implements GroupService {
 
@@ -32,21 +31,33 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public Group insert(Group group) {
+        if (group == null) {
+            throw new GroupException("Group entity can't be empty");
+        }
         return groupDao.insert(group);
     }
 
     @Override
     public void deleteById(Long id) {
+        if (id == null) {
+            throw new GroupException("ID can't be empty");
+        }
         groupDao.deleteById(id);
     }
 
     @Override
-    public Optional<Group> findById(Long id) {
+    public Group findById(Long id) {
+        if (id == null) {
+            throw new GroupException("ID can't be empty");
+        }
         return groupDao.findById(id);
     }
 
     @Override
     public void update(Group group) {
+        if (group == null) {
+            throw new GroupException("Group entity can't be empty");
+        }
         groupDao.update(group);
     }
 
@@ -56,12 +67,31 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public List<Group> findUsersGroups(long id) {
+    public List<Group> findUsersGroups(Long id) {
+        if (id == null) {
+            throw new GroupException("ID can't be empty");
+        }
         return groupDao.findUserGroupsByUserId(id);
     }
 
     @Override
-    public void joinGroup(UserEntity user, Group group) {
-        groupDao.joinGroup(user, group);
+    public void joinGroup(Long userId, Long groupId) {
+        if (userId == null || groupId == null) {
+            throw new GroupException("ID can't be empty");
+        }
+        groupDao.joinGroup(userId, userId);
+    }
+
+    @Override
+    public void leaveGroup(Long userId, Long groupId) {
+        groupDao.leaveGroup(userId, groupId);
+    }
+
+    @Override
+    public boolean checkMembership(Long userId, Long groupId) {
+        if (userId == null || groupId == null) {
+            throw new GroupException("ID can't be empty");
+        }
+        return groupDao.checkMembership(userId, groupId);
     }
 }
